django_server/requirements_prod.txt
Django==4.2.8
djangorestframework==3.14.0
djoser==2.2.2
psycopg2-binary==2.9.1
Pillow==9.0.0
django-cors-headers==4.3.1
reportlab==4.0.8
xhtml2pdf==0.2.13
gunicorn==21.2.0
django-robots-txt==0.4
django-seo-js==0.4.1
django-meta==2.4.1
